fixed regex parsing issue in config.cpp

+++ Makefile
@@ -1,4 +1,4 @@
-CXXFLAGS = -std=c++1z -pthread
+CXXFLAGS = -std=c++1z -pthread -g
 INCLUDES = -I./include
 CURL = -lcurl
 NCURSES = -lncurses
@@ -8,21 +8,21 @@ OBJECTS = dropboxstorage.o csvfilestorage.o wifiinternetdevice.o helperfunctions
 
 OBJ_COMMAND = $(CXX) $(CXXFLAGS) $(INCLUDES) -c $<
 
-.PHONY: all clean install
+.PHONY: all clean install Networking
 
 all: sensorsystem-hub
 
-install:
-	cp ./sensorsystem-hub /usr/local/bin/
-	mkdir ~/.sensorsystem-hub
+install: all
+	install -m 0755 ./sensorsystem-hub /usr/local/bin/
 	
 clean:
-	rm *.o
+	if [ -f *.o ];then rm *.o; fi
 	rm ./sensorsystem-hub
-	rm /usr/local/bin/sensorsystem-hub
+	if [ -f /usr/local/bin/sensorsystem-hub ]; then rm /usr/local/bin/sensorsystem-hub; fi
+	rm -rf ./dep
 
-sensorsystem-hub: $(OBJECTS) ./src/main.cpp
-	$(CXX) $(CXXFLAGS) $(INCLUDES) -o $@ $^ $(FILESYSTEM) $(NETWORKING) $(NCURSES) $(CURL)
+sensorsystem-hub: $(OBJECTS) ./src/main.cpp Networking
+	$(CXX) $(CXXFLAGS) $(INCLUDES) -o $@ $(OBJECTS) ./src/main.cpp $(FILESYSTEM) $(NETWORKING) $(NCURSES) $(CURL)
 
 dropboxstorage.o: ./src/storage/dropboxstorage.cpp ./include/hub/storage/dropboxstorage.h ./include/hub/storage/remotestorage.h
 	$(OBJ_COMMAND)
@@ -39,8 +39,14 @@ helperfunctions.o: ./src/internet/helperfunctions.cpp ./include/hub/internet/hel
 cursesui.o: ./src/cursesui.cpp ./include/hub/cursesui.h
 	$(OBJ_COMMAND)
 	
-config.o: ./src/config.cpp ./include/hub/config.h
+config.o: ./src/config.cpp ./include/hub/config.h Networking
 	$(OBJ_COMMAND)
 	
 measurement.o: ./src/measurement.cpp ./include/hub/measurement.h
 	$(OBJ_COMMAND)
+
+Networking:
+	if [ ! -d ./dep ]; then\
+		git clone https://github.com/tfrec-kalcsits/SensorSystem-Networking.git ./dep/SensorSystem-Networking; \
+		cd ./dep/SensorSystem-Networking && make all && sudo make install; \
+	fi
